user login

+++ app/Http/Controllers/TasksController.php
@@ -15,11 +15,28 @@ class TasksController extends Controller
      */
     public function index()
     {
-        $tasks = Task::all();
+        $data = [];
+        if (\Auth::check()) { // 認証済みの場合
+            // 認証済みユーザを取得
+            $user = \Auth::user();
+            // ユーザの投稿の一覧を作成日時の降順で取得
+            // （後のChapterで他ユーザの投稿も取得するように変更しますが、現時点ではこのユーザの投稿のみ取得します）
+            $tasks = Task::all();
+            //$tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
 
-        return view('tasks.index', [
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+        }
+        
+        
+        // Welcomeビューでそれらを表示
+        return view('tasks.index', $data);
+
+        /*return view('tasks.index', [
             'tasks' => $tasks,
-        ]);
+        ]);*/
     }
 
     /**
@@ -29,6 +46,8 @@ class TasksController extends Controller
      */
     public function create()
     {
+        
+        
         $task = new Task;
 
         return view('tasks.create', [
@@ -48,14 +67,20 @@ class TasksController extends Controller
             'status'=>'required|max:10',
             'content'=>'required'
             ]);
-            
+        /*
         $task = new Task;
         $task->status = $request->status;
         $task->content = $request->content;
         $task->save();
-
-        // トップページへリダイレクトさせる
-        return redirect('/');
+        */
+        // 認証済みユーザ（閲覧者）の投稿として作成（リクエストされた値をもとに作成）
+        $request->user()->tasks()->create([
+            'status' => $request->status,
+            'content' => $request->content,
+            ]);
+    
+        // 前のURLへリダイレクトさせる
+        return back();
     }
 
     /**
@@ -100,6 +125,7 @@ class TasksController extends Controller
         $request->validate([
             'status'=>'required|max:10',
             'content'=>'required'
+            
             ]);
             
         $task = Task::findOrFail($id);
@@ -121,8 +147,11 @@ class TasksController extends Controller
     public function destroy($id)
     {
         $task = Task::findOrFail($id);
-        // 削除
-        $task->delete();
+        
+        // 認証済みユーザ（閲覧者）がその投稿の所有者である場合は、投稿を削除
+        if (\Auth::id() === $task->user_id) {
+            $task->delete();
+        }
 
         // トップページへリダイレクトさせる
         return redirect('/');
